Refactor Ride actions hpp=>h/cpp
