+++ src/openrct2/ReplayManager.cpp
@@ -17,8 +17,8 @@
 #include "PlatformEnvironment.h"
 #include "actions/FootpathPlaceAction.h"
 #include "actions/GameAction.h"
-#include "actions/RideEntranceExitPlaceAction.hpp"
-#include "actions/RideSetSetting.hpp"
+#include "actions/RideEntranceExitPlaceAction.h"
+#include "actions/RideSetSettingAction.h"
 #include "actions/SetCheatAction.hpp"
 #include "actions/TileModifyAction.hpp"
 #include "actions/TrackPlaceAction.hpp"
