+++ src/openrct2/actions/GameActionCompat.cpp
@@ -12,10 +12,10 @@
 #include "MazeSetTrackAction.h"
 #include "PlaceParkEntranceAction.h"
 #include "PlacePeepSpawnAction.h"
-#include "RideCreateAction.hpp"
-#include "RideDemolishAction.hpp"
-#include "RideSetName.hpp"
-#include "RideSetStatus.hpp"
+#include "RideCreateAction.h"
+#include "RideDemolishAction.h"
+#include "RideSetNameAction.h"
+#include "RideSetStatusAction.h"
 #include "SetParkEntranceFeeAction.hpp"
 #include "StaffSetNameAction.hpp"
 #include "WallRemoveAction.hpp"
