+++ src/openrct2/actions/GameActionRegistration.cpp
@@ -50,17 +50,17 @@
 #include "PlacePeepSpawnAction.h"
 #include "PlayerKickAction.h"
 #include "PlayerSetGroupAction.h"
-#include "RideCreateAction.hpp"
-#include "RideDemolishAction.hpp"
-#include "RideEntranceExitPlaceAction.hpp"
-#include "RideEntranceExitRemoveAction.hpp"
-#include "RideSetAppearanceAction.hpp"
-#include "RideSetColourScheme.hpp"
-#include "RideSetName.hpp"
-#include "RideSetPriceAction.hpp"
-#include "RideSetSetting.hpp"
-#include "RideSetStatus.hpp"
-#include "RideSetVehiclesAction.hpp"
+#include "RideCreateAction.h"
+#include "RideDemolishAction.h"
+#include "RideEntranceExitPlaceAction.h"
+#include "RideEntranceExitRemoveAction.h"
+#include "RideSetAppearanceAction.h"
+#include "RideSetColourSchemeAction.h"
+#include "RideSetNameAction.h"
+#include "RideSetPriceAction.h"
+#include "RideSetSettingAction.h"
+#include "RideSetStatusAction.h"
+#include "RideSetVehicleAction.h"
 #include "ScenarioSetSettingAction.hpp"
 #include "SetCheatAction.hpp"
 #include "SetParkEntranceFeeAction.hpp"
