+++ src/openrct2/actions/RideSetPriceAction.cpp
@@ -7,7 +7,7 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
-#pragma once
+#include "RideSetPriceAction.h"
 
 #include "../Cheats.h"
 #include "../common.h"
@@ -21,179 +21,155 @@
 #include "../ride/ShopItem.h"
 #include "../world/Park.h"
 #include "../world/Sprite.h"
-#include "GameAction.h"
 
-DEFINE_GAME_ACTION(RideSetPriceAction, GAME_COMMAND_SET_RIDE_PRICE, GameActions::Result)
+void RideSetPriceAction::AcceptParameters(GameActionParameterVisitor& visitor)
 {
-private:
-    NetworkRideId_t _rideIndex{ RideIdNewNull };
-    money16 _price{ MONEY16_UNDEFINED };
-    bool _primaryPrice{ true };
+    visitor.Visit("ride", _rideIndex);
+    visitor.Visit("price", _price);
+    visitor.Visit("isPrimaryPrice", _primaryPrice);
+}
 
-public:
-    RideSetPriceAction() = default;
-    RideSetPriceAction(ride_id_t rideIndex, money16 price, bool primaryPrice)
-        : _rideIndex(rideIndex)
-        , _price(price)
-        , _primaryPrice(primaryPrice)
+void RideSetPriceAction::Serialise(DataSerialiser& stream)
+{
+    GameAction::Serialise(stream);
+
+    stream << DS_TAG(_rideIndex) << DS_TAG(_price) << DS_TAG(_primaryPrice);
+}
+
+GameActions::Result::Ptr RideSetPriceAction::Query() const
+{
+    GameActions::Result::Ptr res = std::make_unique<GameActions::Result>();
+
+    auto ride = get_ride(_rideIndex);
+    if (ride == nullptr)
     {
+        log_warning("Invalid game command, ride_id = %u", uint32_t(_rideIndex));
+        return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
     }
 
-    void AcceptParameters(GameActionParameterVisitor & visitor) override
+    rct_ride_entry* rideEntry = get_ride_entry(ride->subtype);
+    if (rideEntry == nullptr)
     {
-        visitor.Visit("ride", _rideIndex);
-        visitor.Visit("price", _price);
-        visitor.Visit("isPrimaryPrice", _primaryPrice);
+        log_warning("Invalid game command for ride %u", uint32_t(_rideIndex));
+        return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
     }
 
-    uint16_t GetActionFlags() const override
+    return res;
+}
+
+GameActions::Result::Ptr RideSetPriceAction::Execute() const
+{
+    GameActions::Result::Ptr res = std::make_unique<GameActions::Result>();
+    res->Expenditure = ExpenditureType::ParkRideTickets;
+
+    auto ride = get_ride(_rideIndex);
+    if (ride == nullptr)
     {
-        return GameAction::GetActionFlags() | GameActions::Flags::AllowWhilePaused;
+        log_warning("Invalid game command, ride_id = %u", uint32_t(_rideIndex));
+        return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
     }
 
-    void Serialise(DataSerialiser & stream) override
+    rct_ride_entry* rideEntry = get_ride_entry(ride->subtype);
+    if (rideEntry == nullptr)
     {
-        GameAction::Serialise(stream);
-
-        stream << DS_TAG(_rideIndex) << DS_TAG(_price) << DS_TAG(_primaryPrice);
+        log_warning("Invalid game command for ride %u", uint32_t(_rideIndex));
+        return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
     }
 
-    GameActions::Result::Ptr Query() const override
+    if (!ride->overall_view.isNull())
     {
-        GameActions::Result::Ptr res = std::make_unique<GameActions::Result>();
-
-        auto ride = get_ride(_rideIndex);
-        if (ride == nullptr)
-        {
-            log_warning("Invalid game command, ride_id = %u", uint32_t(_rideIndex));
-            return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
-        }
-
-        rct_ride_entry* rideEntry = get_ride_entry(ride->subtype);
-        if (rideEntry == nullptr)
-        {
-            log_warning("Invalid game command for ride %u", uint32_t(_rideIndex));
-            return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
-        }
-
-        return res;
+        auto location = ride->overall_view.ToTileCentre();
+        res->Position = { location, tile_element_height(location) };
     }
 
-    GameActions::Result::Ptr Execute() const override
+    ShopItem shopItem;
+    if (_primaryPrice)
     {
-        GameActions::Result::Ptr res = std::make_unique<GameActions::Result>();
-        res->Expenditure = ExpenditureType::ParkRideTickets;
-
-        auto ride = get_ride(_rideIndex);
-        if (ride == nullptr)
+        shopItem = ShopItem::Admission;
+        if (ride->type != RIDE_TYPE_TOILETS)
         {
-            log_warning("Invalid game command, ride_id = %u", uint32_t(_rideIndex));
-            return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
-        }
-
-        rct_ride_entry* rideEntry = get_ride_entry(ride->subtype);
-        if (rideEntry == nullptr)
-        {
-            log_warning("Invalid game command for ride %u", uint32_t(_rideIndex));
-            return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
-        }
-
-        if (!ride->overall_view.isNull())
-        {
-            auto location = ride->overall_view.ToTileCentre();
-            res->Position = { location, tile_element_height(location) };
-        }
-
-        ShopItem shopItem;
-        if (_primaryPrice)
-        {
-            shopItem = ShopItem::Admission;
-            if (ride->type != RIDE_TYPE_TOILETS)
-            {
-                shopItem = rideEntry->shop_item[0];
-                if (shopItem == ShopItem::None)
-                {
-                    ride->price[0] = _price;
-                    window_invalidate_by_class(WC_RIDE);
-                    return res;
-                }
-            }
-            // Check same price in park flags
-            if (!shop_item_has_common_price(shopItem))
+            shopItem = rideEntry->shop_item[0];
+            if (shopItem == ShopItem::None)
             {
                 ride->price[0] = _price;
                 window_invalidate_by_class(WC_RIDE);
                 return res;
             }
         }
-        else
+        // Check same price in park flags
+        if (!shop_item_has_common_price(shopItem))
         {
-            shopItem = rideEntry->shop_item[1];
-            if (shopItem == ShopItem::None)
-            {
-                shopItem = RideTypeDescriptors[ride->type].PhotoItem;
-                if ((ride->lifecycle_flags & RIDE_LIFECYCLE_ON_RIDE_PHOTO) == 0)
-                {
-                    ride->price[1] = _price;
-                    window_invalidate_by_class(WC_RIDE);
-                    return res;
-                }
-            }
-            // Check same price in park flags
-            if (!shop_item_has_common_price(shopItem))
+            ride->price[0] = _price;
+            window_invalidate_by_class(WC_RIDE);
+            return res;
+        }
+    }
+    else
+    {
+        shopItem = rideEntry->shop_item[1];
+        if (shopItem == ShopItem::None)
+        {
+            shopItem = RideTypeDescriptors[ride->type].PhotoItem;
+            if ((ride->lifecycle_flags & RIDE_LIFECYCLE_ON_RIDE_PHOTO) == 0)
             {
                 ride->price[1] = _price;
                 window_invalidate_by_class(WC_RIDE);
                 return res;
             }
         }
-
-        // Synchronize prices if enabled.
-        RideSetCommonPrice(shopItem);
-
-        return res;
-    }
-
-private:
-    void RideSetCommonPrice(ShopItem shopItem) const
-    {
-        for (auto& ride : GetRideManager())
+        // Check same price in park flags
+        if (!shop_item_has_common_price(shopItem))
         {
-            auto invalidate = false;
-            auto rideEntry = get_ride_entry(ride.subtype);
-            if (ride.type == RIDE_TYPE_TOILETS && shopItem == ShopItem::Admission)
-            {
-                if (ride.price[0] != _price)
-                {
-                    ride.price[0] = _price;
-                    invalidate = true;
-                }
-            }
-            else if (rideEntry != nullptr && rideEntry->shop_item[0] == shopItem)
-            {
-                if (ride.price[0] != _price)
-                {
-                    ride.price[0] = _price;
-                    invalidate = true;
-                }
-            }
-            if (rideEntry != nullptr)
-            {
-                // If the shop item is the same or an on-ride photo
-                if (rideEntry->shop_item[1] == shopItem
-                    || (rideEntry->shop_item[1] == ShopItem::None && GetShopItemDescriptor(shopItem).IsPhoto()))
-                {
-                    if (ride.price[1] != _price)
-                    {
-                        ride.price[1] = _price;
-                        invalidate = true;
-                    }
-                }
-            }
-            if (invalidate)
-            {
-                window_invalidate_by_number(WC_RIDE, ride.id);
-            }
+            ride->price[1] = _price;
+            window_invalidate_by_class(WC_RIDE);
+            return res;
         }
     }
-};
+
+    // Synchronize prices if enabled.
+    RideSetCommonPrice(shopItem);
+
+    return res;
+}
+
+void RideSetPriceAction::RideSetCommonPrice(ShopItem shopItem) const
+{
+    for (auto& ride : GetRideManager())
+    {
+        auto invalidate = false;
+        auto rideEntry = get_ride_entry(ride.subtype);
+        if (ride.type == RIDE_TYPE_TOILETS && shopItem == ShopItem::Admission)
+        {
+            if (ride.price[0] != _price)
+            {
+                ride.price[0] = _price;
+                invalidate = true;
+            }
+        }
+        else if (rideEntry != nullptr && rideEntry->shop_item[0] == shopItem)
+        {
+            if (ride.price[0] != _price)
+            {
+                ride.price[0] = _price;
+                invalidate = true;
+            }
+        }
+        if (rideEntry != nullptr)
+        {
+            // If the shop item is the same or an on-ride photo
+            if (rideEntry->shop_item[1] == shopItem
+                || (rideEntry->shop_item[1] == ShopItem::None && GetShopItemDescriptor(shopItem).IsPhoto()))
+            {
+                if (ride.price[1] != _price)
+                {
+                    ride.price[1] = _price;
+                    invalidate = true;
+                }
+            }
+        }
+        if (invalidate)
+        {
+            window_invalidate_by_number(WC_RIDE, ride.id);
+        }
+    }
+}
