+++ src/openrct2/interface/InteractiveConsole.cpp
@@ -17,8 +17,8 @@
 #include "../ReplayManager.h"
 #include "../Version.h"
 #include "../actions/ClimateSetAction.h"
-#include "../actions/RideSetPriceAction.hpp"
-#include "../actions/RideSetSetting.hpp"
+#include "../actions/RideSetPriceAction.h"
+#include "../actions/RideSetSettingAction.h"
 #include "../actions/SetCheatAction.hpp"
 #include "../actions/StaffSetCostumeAction.hpp"
 #include "../config/Config.h"
