+++ src/openrct2/libopenrct2.vcxproj
@@ -95,17 +95,17 @@
     <ClInclude Include="actions\PlacePeepSpawnAction.h" />
     <ClInclude Include="actions\PlayerKickAction.h" />
     <ClInclude Include="actions\PlayerSetGroupAction.h" />
-    <ClInclude Include="actions\RideCreateAction.hpp" />
-    <ClInclude Include="actions\RideDemolishAction.hpp" />
-    <ClInclude Include="actions\RideEntranceExitPlaceAction.hpp" />
-    <ClInclude Include="actions\RideEntranceExitRemoveAction.hpp" />
-    <ClInclude Include="actions\RideSetAppearanceAction.hpp" />
-    <ClInclude Include="actions\RideSetColourScheme.hpp" />
-    <ClInclude Include="actions\RideSetName.hpp" />
-    <ClInclude Include="actions\RideSetPriceAction.hpp" />
-    <ClInclude Include="actions\RideSetSetting.hpp" />
-    <ClInclude Include="actions\RideSetStatus.hpp" />
-    <ClInclude Include="actions\RideSetVehiclesAction.hpp" />
+    <ClInclude Include="actions\RideCreateAction.h" />
+    <ClInclude Include="actions\RideDemolishAction.h" />
+    <ClInclude Include="actions\RideEntranceExitPlaceAction.h" />
+    <ClInclude Include="actions\RideEntranceExitRemoveAction.h" />
+    <ClInclude Include="actions\RideSetAppearanceAction.h" />
+    <ClInclude Include="actions\RideSetColourSchemeAction.h" />
+    <ClInclude Include="actions\RideSetNameAction.h" />
+    <ClInclude Include="actions\RideSetPriceAction.h" />
+    <ClInclude Include="actions\RideSetSettingAction.h" />
+    <ClInclude Include="actions\RideSetStatusAction.h" />
+    <ClInclude Include="actions\RideSetVehicleAction.h" />
     <ClInclude Include="actions\ScenarioSetSettingAction.hpp" />
     <ClInclude Include="actions\SetCheatAction.hpp" />
     <ClInclude Include="actions\SetParkEntranceFeeAction.hpp" />
