+++ src/openrct2/ride/Ride.cpp
@@ -15,10 +15,10 @@
 #include "../Game.h"
 #include "../Input.h"
 #include "../OpenRCT2.h"
-#include "../actions/RideEntranceExitRemoveAction.hpp"
-#include "../actions/RideSetSetting.hpp"
-#include "../actions/RideSetStatus.hpp"
-#include "../actions/RideSetVehiclesAction.hpp"
+#include "../actions/RideEntranceExitRemoveAction.h"
+#include "../actions/RideSetSettingAction.h"
+#include "../actions/RideSetStatusAction.h"
+#include "../actions/RideSetVehicleAction.h"
 #include "../actions/TrackRemoveAction.hpp"
 #include "../audio/AudioMixer.h"
 #include "../audio/audio.h"
