+++ src/openrct2/ride/TrackDesign.cpp
@@ -18,11 +18,11 @@
 #include "../actions/LargeSceneryPlaceAction.h"
 #include "../actions/LargeSceneryRemoveAction.h"
 #include "../actions/MazePlaceTrackAction.h"
-#include "../actions/RideCreateAction.hpp"
-#include "../actions/RideEntranceExitPlaceAction.hpp"
-#include "../actions/RideSetName.hpp"
-#include "../actions/RideSetSetting.hpp"
-#include "../actions/RideSetVehiclesAction.hpp"
+#include "../actions/RideCreateAction.h"
+#include "../actions/RideEntranceExitPlaceAction.h"
+#include "../actions/RideSetNameAction.h"
+#include "../actions/RideSetSettingAction.h"
+#include "../actions/RideSetVehicleAction.h"
 #include "../actions/SmallSceneryPlaceAction.h"
 #include "../actions/SmallSceneryRemoveAction.h"
 #include "../actions/TrackPlaceAction.hpp"
