+++ src/openrct2/world/Entrance.cpp
@@ -13,8 +13,8 @@
 #include "../Game.h"
 #include "../OpenRCT2.h"
 #include "../actions/ParkEntranceRemoveAction.h"
-#include "../actions/RideEntranceExitPlaceAction.hpp"
-#include "../actions/RideEntranceExitRemoveAction.hpp"
+#include "../actions/RideEntranceExitPlaceAction.h"
+#include "../actions/RideEntranceExitRemoveAction.h"
 #include "../localisation/StringIds.h"
 #include "../management/Finance.h"
 #include "../network/network.h"
